1006A adjacent replacements
Mishka got an integer array a of length n as a birthday present (what a surprise!).

Mishka doesn't like this present and wants to change it somehow. He has invented an algorithm and called it "Mishka's Adjacent Replacements Algorithm". This algorithm can be represented as a sequence of steps:

Replace each occurrence of 1 in the array a with 2;
Replace each occurrence of 2 in the array a with 1;
Replace each occurrence of 3 in the array a with 4;
Replace each occurrence of 4 in the array a with 3;
Replace each occurrence of 5 in the array a with 6;
Replace each occurrence of 6 in the array a with 5;
…
Replace each occurrence of 109−1 in the array a with 109;
Replace each occurrence of 109 in the array a with 109−1.